% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic.helper.fun.R
\name{calc.ssmd}
\alias{calc.ssmd}
\title{Calculate SSMDs given a df (or equivalent) of means and stds}
\usage{
calc.ssmd(mdata, mbase, sddata, sdbase, scenario)
}
\arguments{
\item{mdata}{Data containing mean values}

\item{mbase}{Data containing mean values for base scenario}

\item{sddata}{Data containing standard deviation values}

\item{sdbase}{Data containing standard deviation values for base scenario}

\item{scenario}{The name of the scenario for which SSMDs are being calculated}
}
\description{
Calculate SSMDs given a df (or equivalent) of means and stds.
}
\details{
Checks whether the two dfs have same column names and attempt to re-order the 
  columns if not.
}
