% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/HexSimR.R
\name{census.calc}
\alias{census.calc}
\title{Carry out calculations on a subset of variables in a specific census file}
\usage{
census.calc(path.results = NULL, ncensus, headers, var.name = NULL,
  bin.f = "+", scenarios = "all")
}
\arguments{
\item{path.results}{The path for the 'Results' folder}

\item{ncensus}{The census number to be processed}

\item{headers}{A character vector matching exactly the headers of the census
file columns that are to be used by \code{bin.f}}

\item{var.name}{A character vector to name the new variable. Mandatory if
\code{bin.f} is of class \code{function}. See details.}

\item{bin.f}{A character vector to pass the binary function to be applied to
the columns identified by \code{headers} (Default: \code{"+"})}

\item{scenarios}{A character vector with scenarios to be processed or "all"}
}
\value{
Save census file to disk and return a list with the new census files
}
\description{
\code{census.calc} carries out calculations indicated with the argument \code{bin.f}.
  For example, this will be needed when trait are sub-classes and there is
  the need to sum a subset of them (e.g. traits are divided by gender and age
  and the user wants to have the total by gender) or calculate a proportion
  (e.g. sex-ratio or over the total). Therefore, typically, these operations
  will be sums, subtractions, divisions or
  multiplications. However, \code{bin.f} can take any function that it is
  compatible with \code{Reduce()} (default is \code{"+"}).
}
\details{
The existing census file will be overwritten with the new census file, which
  will include the new variable.

The subset of data to be used by \code{bin.f} is indicated with \code{headers}.
  This argument takes a character vector with the headers of the census files.
  These have to match exactly the headers in the census file (including spaces).
  \code{headers} is internally modified with \code{make.names()} to guarantee
  that names are syntactically valid. Therefore, a character vector where the
  headers were already modified and would match the results of \code{make.names()}
  is also acceptable.

The argument \code{var.name} is the name of the new variable being calculated.
  If none is passed, then, by default, \code{census.calc} will paste together
  the names passed with \code{headers} separated by the operator passed with
  \code{bin.f}. If \code{bin.f} is a function (i.e. a R object of class
  'function'), the default behaviour will fail and an error is reported. In
  this case, the user have to pass a valid \code{var.name}. It is recommended
  to use letters, numbers and the dot or underline ("_") characters, to start with
  a letter or the dot not followed by a number.

If more then one calculations needs to be performed, of the same calculations
  needs to be performed on different groups of variables (headers), then
  \code{census.calc} needs to be run multiple times.
}

