% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{census.plot}
\alias{census.plot}
\title{Plots census means across simulated time steps}
\usage{
census.plot(path.results = NULL, scenarios = "all", traits,
  ncensus = 0, ngroups = 1, rm.T0 = TRUE)
}
\arguments{
\item{path.results}{The path to the 'Results' folder}

\item{scenarios}{A character vector with the scenarios to be processed or
"all"}

\item{traits}{A character vector with the name of the traits to be included}

\item{ncensus}{The number of the census to be considered}

\item{ngroups}{The number of groups in which the scenarios are to be divided}

\item{rm.T0}{Whether to remove the first row of data. Relevant when it is 
Time Step '0' (default=TRUE)}
}
\value{
Save to disk ggplot objects (with extension .rda) and pdf with the 
  plots (one for each group). Return a list of plots
}
\description{
It takes as data input the output from \code{collate.census} (it reads data 
directly from xls files).
}
\details{
\code{rm.T0} is generally used to remove the the first row of data. This
makes sense when all time steps were logged, but there is little meaning in 
keeping in time zero as this is the initialization step.
}
