% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genepop.utilities.R
\name{clean.genepop}
\alias{clean.genepop}
\title{Remove illegal characters from HexSim generated genepop input file}
\usage{
clean.genepop(fname, title = NULL)
}
\arguments{
\item{fname}{A character vector with the name of the input file including the 
path}

\item{title}{A character vector to be used to replace the first line in the file}
}
\value{
Save to disk a cleaned up genepop input file and a character vector
}
\description{
HexSim includes a 'structure block' at the end of its genepop input file
  that are not recognised by genepop. This block is removed by \code{clean.genepop}
  Also, while the animal IDs between brackets should not create a problem in
  genepop, these may be an 'unexpected column' for other applications that 
  reads genepop input files (e.g. the R package diveRsity). Similarly, the 
  space between the word 'Trait' and the trait number, and the equal and column
  symbols in the first line of the file can cause unexpected 
  behaviours in some applications. All these unusual features are removed from
  the HexSim generated files, which are then re-saved with the same name,  
  a suffix "cleaned" and an extensin ".gen".
}

