% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/copy.results.R
\name{copy.results}
\alias{copy.results}
\title{Copy \code{HexSimR} results}
\usage{
copy.results(path.results, out = getwd(), copy.invasion = TRUE,
  inv.name = "Invasion.front.xlsx", Pext = TRUE, comp.census = TRUE,
  comp.move = TRUE, move.name = "SSMD_move.xlsx", comp.ranges = TRUE,
  ranges.name = "SSMD_ranges.xlsx", plots = TRUE,
  scen.results = TRUE, scenarios = "all")
}
\arguments{
\item{path.results}{The path to the 'Results' folder}

\item{out}{The directory where to save the results}

\item{copy.invasion}{Whether to copy data saved from \code{invasion.front}}

\item{inv.name}{The name of file with results from   \code{invasion.front}}

\item{Pext}{Whether to save results from \code{Pext}}

\item{comp.census}{Whether to copy data saved from \code{SSMD_census}}

\item{comp.move}{Whether to copy data saved from \code{SSMD_move}}

\item{move.name}{The name of file with results from   \code{SSMD_move}}

\item{comp.ranges}{Whether to copy data saved from \code{SSMD_ranges}}

\item{ranges.name}{The name of file with results from   \code{SSMD_ranges}}

\item{plots}{Whether to copy plots}

\item{scen.results}{Whether to copy results saved in scenarios' folders}

\item{scenarios}{A character vector with the scenarios to be processed or 
"all"}
}
\description{
\strong{This function is deprecated. Please see '?compress.log' to reduce project's size}
}
\details{
This function is used to copy/back up  \code{HexSimR} results maintaining the
 same directory structure.
  
\code{copy.results} will duplicate all folders present within the 'Results' 
  folder when \code{scen.results=TRUE}. Note that only the first level of 
  results will be copied (i.e. no individual replicate results are copied). 
  All .xlsx files present in scenarios' folders will be copied (even 
  those with manually modified names) assuming that these are results to keep. 
  These include genetic mean distances.

It assummes that file names of plots start with "plot" as for   \code{HexSimR} 
  default. If exists, the file "gen.plot.data.csv" (if the name was not 
  modified) will be copied as part of this option because considered part of 
  the plot files.
  
With \code{comp.census}, \code{comp.move}, and \code{comp.ranges} can be set 
  whether to copy results of statistical comparisons (output from \code{SSMD_census},
  \code{SSMD_move}, \code{SSMD_ranges}).
  
The SSMD file generated with \code{Pext} is copied automatically (if exists 
   and if the name was not modified), along with the other results from this
   function, when the option \code{Pext=TRUE} is selected.
  
When \code{comp.census=TRUE}, all files whose name starts with "SSMD_census"
   will be copied (note that if some file names has been modified and does  
   not start with "SSMD_census" anymore, these files will not be copied).
}
