% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compress.logs.R
\name{decompress.gz}
\alias{decompress.gz}
\title{Decompress files with ".gz" extension, append ".log" extension and optinally
remove the .gz files}
\usage{
decompress.gz(path.results = NULL, scenarios = "all",
  delete.gz = TRUE, overwrite.log = TRUE, parallel = FALSE,
  ncores = NULL)
}
\arguments{
\item{path.results}{The path to the 'Results' folder}

\item{scenarios}{A character vector with the scenarios to be processed or
"all"}

\item{delete.gz}{whether the .gz file should be deleted after decompression}

\item{overwrite.log}{if a .log file with the same name is present, should it be
overwritten?}

\item{parallel}{whether execute the function in parallel}

\item{ncores}{the number of cores if executed in parallel}
}
\value{
a list where the first element is the list of gz files found and the
  second is the log files created
}
\description{
Decompress files with ".gz" extension, append ".log" extension and optinally
remove the .gz files, which may be necessay if reports need to be created
}
\details{
Note that it makes sense to execute this function in parallel only if a solid
state disk (SSD) is available
}
