% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genepop.utilities.R
\name{gen.dist}
\alias{gen.dist}
\title{Genetic distance}
\usage{
gen.dist(fname, mean.type = "harmonic")
}
\arguments{
\item{fname}{A character vector with the name of the input file including the 
path}

\item{mean.type}{The type of mean to be calculated over multiple loci}
}
\value{
Save to disk a file with same root of the input file with extension 
  .pairD
}
\description{
\code{gen.dist} calculates the Jost's D between all possible population pairs
   using the R package \code{mmod}.
   
This function takes the output from the function \code{clean.genepop} 
  (or \code{multi.clean.genepop}) as input and uses the R package \code{adegenet}
  to convert the input in a genind object. 
    
A text file with extension .pairD that contains a pairwise distance  matrix 
  is saved to disk.
  
\code{mean.type} can be set to either "arithmetic" or "harmonic" (default), 
  which is used to calculate the global estimates of Hs and Ht (see 
  \code{mmod}'s manual for details).
}
\references{
Jost, L.O.U., 2008. GST and its relatives do not measure 
  differentiation. Molecular Ecology 17, 4015-4026.

Winter, D.J., 2012. mmod: an R library for the calculation of 
  population differentiation statistics. Molecular Ecology Resources 12, 
  1158-1160.
}

