% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genepop.div.utilities.R
\name{gen.div}
\alias{gen.div}
\title{Genetic Diversity}
\usage{
gen.div(gi)
}
\arguments{
\item{gi}{A genind object}
}
\value{
The output from gl.report.heterozygosity
}
\description{
\code{gen.div} calculates genetic diversity measurements
   using the R package \code{dartRverse}.
   
This function takes as input a genind object
}
