% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{gen.plot}
\alias{gen.plot}
\title{Plot genetic distance with standard deviation bars}
\usage{
gen.plot(path.results = NULL, scenarios = "all", pop.name = NULL,
  time.step = 1, traits, r = 2, c = 1)
}
\arguments{
\item{path.results}{The path to the 'Results' folder}

\item{scenarios}{A character vector with the scenarios to be processed or "all"}

\item{time.step}{A numeric vector of length 1 to indicate the time step to 
be included}

\item{r}{The raw in the genetic distance matrix to use}

\item{c}{The column in the genetic distance matrix to use}
}
\description{
\code{gen.plot} generates a plot of the mean genetic distances with  
  standard deviation bars. It assumes that file names are as for \code{HexSimR}
  defaults.
}

