\name{HexSimR-package}
\alias{HexSimR-package}
\alias{HexSimR}
\docType{package}
\title{
\packageTitle{HexSimR}
}
\description{
\packageDescription{HexSimR}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{HexSimR}
\packageIndices{HexSimR}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{HexSimR}

Maintainer: \packageMaintainer{HexSimR}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
