% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/invasion.front.R
\name{invasion.front}
\alias{invasion.front}
\title{Calculate the progressive occupancy of an array of hexagons}
\usage{
invasion.front(path.results = NULL, ncensus, value = 1, patch.width,
  scenarios = "all")
}
\arguments{
\item{path.results}{The path to the 'Results' folder}

\item{ncensus}{The census number to be processed}

\item{value}{The threshold occupancy value}

\item{patch.width}{The width of the patch}

\item{scenarios}{A character vector with the scenarios to be processed or "all"}
}
\value{
A list with three elements: the mean and standard
  deviation for each time step and overall.
}
\description{
The user has to set up an array of patches (which may comprise of one or more 
  hexagons) and then determine the occupancy 
  of these patches during the simulations. This can be achieved by generating 
  a sequenced trait linked to an accumulator that it is updated with an 
  'individual location' updater. Finally, by setting up a census event that 
  summarises the number of individuals within each patch, it will be 
  possible to use this function to calculate the progress of the front of 
  invasion across the array of hexagons.
  
The census file where the data is stored is identified with the argument 
  \code{ncensus}. Remember that by default HexSim names the census files with
  the name of the scenario, followed by the 'number' of the census event. This
  'number' will depend on how many census event you have in the HexSim sequence.
  The first census event will be saved with the number 0, the second with the
  number 1 and so forth. The number of the census file that holds the individual
  location data is the one that need to be passed to \code{ncensus} so that 
  \code{HexSimR} knows which one to pick to carry out the calculations.
}

