% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plots.R
\name{invasion.plot}
\alias{invasion.plot}
\title{Plots mean progress of invasion front}
\usage{
invasion.plot(fname = NULL)
}
\arguments{
\item{fname}{A character vector with the name of the input file including the
path}
}
\value{
Save to disk ggplot objects (with extension .rda), a pdf with the
  plot and return the plot
}
\description{
\code{invasion.plot} plots the mean progress of invasion front for each
  scenario included in the output of \code{invasion.front} (it reads data
  directly from xls files).
}

