% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xml_utilities.R
\name{is.attribs}
\alias{is.attribs}
\title{Check whether the param is an attribute or a node}
\usage{
is.attribs(
  attrib,
  param_node_attribute,
  param_node_identifier,
  param_identifier
)
}
\arguments{
\item{attrib}{Whether the identifier is an attribute}

\item{param_node_attribute}{Whether the param_node identifier is an attribute}

\item{param_node_identifier}{Whether param_node has an identifier}

\item{param_identifier}{Whether the param has an identifier (it is assumed
this is an attribute)}
}
\value{
A logical vector with the Xpathof length=1
}
\description{
Check whether the param being changed by LHS.scenarios or 
xml.cond.replacement is an attribute or a node. This function is used 
internally by HexSimR.
}
