% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic.helper.fun.R
\name{iter.folders}
\alias{iter.folders}
\title{Detect iteration folders}
\usage{
iter.folders(dir.path, scenario)
}
\arguments{
\item{dir.path}{The path where the 'Results' folder is located}

\item{scenario}{Scenario to be processed. This is normally passed with lapply()}
}
\value{
A list where each element is the path to the replicates of a scenario
}
\description{
This function is used by a number of HexSimR functions to create a list of 
  path where results for each iteration are saved.
}
