% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xml_utilities.R
\name{LHS.scenarios}
\alias{LHS.scenarios}
\title{Generate LHS scenarios}
\usage{
LHS.scenarios(path.scenarios = NULL, xml.template = NULL, samples, csv.in,
  generate = TRUE)
}
\arguments{
\item{path.scenarios}{The path to the 'Scenarios' folder}

\item{xml.template}{The name of the xml file to use as template for the new 
nodes}

\item{samples}{The number of LHS samples (i.e. parameter combinations)}

\item{csv.in}{The name of the .csv file in the "Scenarios" folder}

\item{generate}{Whether generate (TRUE) the xml files or stop after having
reated the hypercube matrix (FALSE)}
}
\description{
Generate LHS scenarios
}
