% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic.helper.fun.R
\name{make.barrier}
\alias{make.barrier}
\title{Modify a hbf with HexSim barrier data}
\usage{
make.barrier(template, mortality, deflection, npairs, file.name, sufs = NULL,
  dir.out = NULL)
}
\arguments{
\item{template}{The full name of the hbf barrier file to use as template}

\item{mortality}{is the vector of new mortality values that are replaced}

\item{deflection}{is the vector of new deflection values that are replaced}

\item{npairs}{The number of barrier section}

\item{file.name}{A character vector with the output file name}

\item{sufs}{A character vector with suffix(es)}

\item{dir.out}{Path where to save the output. If NULL (default), the working 
directory is used}
}
\description{
This function replaces mortality and deflection values and saves the file 
with a new name. Mortality and deflection values have to be passed in pairs 
for each fence section. So, if you want to generate three files, and you only
have one barrier pair you have to provide six values.
}
\details{
Several files can be generated using mapply() to pass one \code{template} and
one \code{file.name} at the time.
}
