% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic.helper.fun.R
\name{make.map}
\alias{make.map}
\title{Modify a csv with hexmap data}
\usage{
make.map(
  template,
  new.values,
  old.value,
  file.name,
  sufs = NULL,
  dir.out = NULL
)
}
\arguments{
\item{template}{The full name of the csv map file to use as template}

\item{new.values}{is the vector of values that are replaced}

\item{old.value}{is the value that is searched and replaced in the csv file}

\item{file.name}{A character vector with the output file name}

\item{sufs}{A character vector with suffix(es)}

\item{dir.out}{Path where to save the output. If NULL (default), the working 
directory is used}
}
\description{
This function searches for a value and replaces it with \code{new.values}. When
\code{length(new.values) > 1}, \code{sufs} needs to be passed to append a 
suffix to the file name, otherwise the file is overwritten.
}
