% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xml_utilities.R
\name{make.Xpath}
\alias{make.Xpath}
\title{Buil a Xpath}
\usage{
make.Xpath(node_path, identifier, attrib, param_node = NA, is.LHS = FALSE)
}
\arguments{
\item{node_path}{The node_path}

\item{identifier}{The identifier of the node}

\item{attrib}{Whether the identifier is an attribute}
}
\description{
This function builds a Xpath to search nodes in HexSim xml files. Is used 
internally by hexSimR. When used in the LHS.scenarios function, it constructs
the path "/.../name[text()='...']" when param_node is not NA and is.LHS is
TRUE. Addtional details on params are in the scenarios.batch.modifier.
}
