% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xml_utilities.R
\name{make.Xpaths}
\alias{make.Xpaths}
\title{Check conditions and pass the right arguments to make.Xpath}
\usage{
make.Xpaths(
  node_paths,
  Xpaths,
  identifiers,
  attribs,
  param_nodes = NA,
  param_node_identifiers = FALSE,
  param_node_attributes = FALSE,
  param_identifiers = FALSE,
  is.LHS = FALSE
)
}
\arguments{
\item{node_paths}{A character vector of node_paths}

\item{Xpaths}{A empty list of length = length(node_paths)}

\item{identifiers}{A character vector of node identifiers}

\item{attribs}{A logical vector as to whether the identifiers are attributes}

\item{param_nodes}{A character vector if the parameters are contained in
internal nodes, NA otherwise}

\item{param_node_identifiers}{A character vector if the param_nodes have
identifiers, FALSE otherwise}

\item{param_node_attributes}{Whether the param_node_identifiers is an
attribute (TRUE) or not (FALSE).}

\item{param_identifiers}{A character vector if the params have identifiers
(it is assumed these are attributes), FALSE otherwise}

\item{is.LHS}{A logical vector whether the Xpaths being built are for a LHS}
}
\value{
A list (Xpaths) of character vectors with the Xpaths
}
\description{
See scenarios.batch.modifier for addtional details on arguments
}
