% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HexSimR.R
\name{move}
\alias{move}
\title{Calculates descriptive stats from the HexSim generated report 'movement'}
\usage{
move(rep.move = NULL)
}
\arguments{
\item{rep.move}{The fully qualified (i.e. including the path) name of the report 
'movement'}
}
\value{
A data.frame where each column is a HexSim's Event (and it is saved 
  to disk as .csv file)
}
\description{
Calculates descriptive stats from the HexSim generated report 'movement'
}

