% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genepop.utilities.R
\name{multi.clean.genepop}
\alias{multi.clean.genepop}
\title{Remove illegal characters from several HexSim generated genepop input files}
\usage{
multi.clean.genepop(path.results = NULL, scenarios = "all",
  pop.name = NULL)
}
\arguments{
\item{path.results}{The path to the 'Results' folder}

\item{scenarios}{A character vector with the scenarios to be processed or "all"}

\item{pop.name}{The name of the population}
}
\value{
Save to disk a genepop file with a suffix 'cleaned' and an extensin
     ".gen" for each iteration within eachscenario
}
\description{
\code{multi.clean.genepop} is a wrapper for \code{clean.genepop} that processes
  all HexSim generated genepop files within a given scenario(s). It assumes 
  that the names of the report files were 
  not changed from defaults. It reads the files in 
  each of the scenario's iteration subfolder and processes them   
  using \code{clean.genepop}.
  
\code{multi.clean.genepop} identifies the files searching for a text file 
  whose root's name is the \emph{scenario_name} of the file and the text 
  'REPORT_genepop', so if you have modified the name of the files, this 
  function won't work. Similarly, if you have added files named with a pattern
  consistent with what described above, the function will try to process these
  as well.   
  
When using \code{multi.clean.genepop} there is no option to choose the title 
  of the genepop files 
  (i.e. the first line of the genepop input file).
}
\seealso{
clean.genepop
}

