% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/genepop.utilities.R
\name{multi.gen.dist}
\alias{multi.gen.dist}
\title{Genetic distance of multiple scenarios}
\usage{
multi.gen.dist(path.results = NULL, scenarios = "all", pop.name = NULL,
  mean.type = "harmonic")
}
\arguments{
\item{path.results}{The path for the 'Results' folder}

\item{scenarios}{A character vector with scenarios to be processed or "all"}

\item{pop.name}{The name of the population}

\item{mean.type}{The type of mean to be calculated over multiple loci}
}
\value{
See \code{gen.dist} for each scenario
}
\description{
\code{multi.gen.dist} is a wrapper for \code{gen.dist} that generated
  a matrix of genetic distanced for multiple scenarios.
}
\details{
It will generate a matrix for each replicate within the scenario's folder.
}
\seealso{
gen.dist
}

