% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HexSimR.R
\name{multi.reports}
\alias{multi.reports}
\title{Generate summary statistics for multiple reports}
\usage{
multi.reports(path.results = NULL, scenarios = "all", pop.name = NULL,
  type = "move", all = TRUE, hx = NULL, events = NULL, start = "min",
  end = "max")
}
\description{
\code{multi.reports} is basically a wrapper for the two functions that process
  HexSim generated reports. It assumes that the names of the report files were 
  not changed from defaults and that each report was generated in a similar 
  fashion. That is, either they are all generated from a combined log file, or
  they are generated from each iteration's log file. It reads the reports in 
  each of the scenario's folder (or iteration subfolders) and process them as  
  using either \code{move} or \code{ranges}.
  
If the reports file is generated from each iteration's log file, then the 
  reports are sitting within each iteration's subfolder and \code{path.results} 
  has to point to the scenario's folder (i.e. the subfolders will be the 
  scenario's name with the suffix [x], where x is the number of the iteration).
  
  @param pop.name The name of the population
  @param type Whether movement ("move") or ranges ("ranges") reports should be 
    processed
  @param all Whether the reports were generated from a combined log file
  @inheritParams SSMD.census
  @inheritParams ranges
  @return A list where each element is the output from either \code{move} or
    \code{ranges}.
  @export
}

