% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Pext.R
\name{Pext}
\alias{Pext}
\title{Probability of extinction}
\usage{
Pext(data = NULL, path.results, rda.in = "collated.census.rda",
  scenarios = "all", ncensus = 0, start = "min", end = "max",
  headers = "Population.Size", base = NULL)
}
\arguments{
\item{data}{The output from \code{collate.census}}

\item{path.results}{The path to the 'Results' folder}

\item{rda.in}{The fully qualified (i.e. including the path) name of the .rda 
file where the output from \code{collate.census} is saved}

\item{scenarios}{A character vector with the scenarios to be processed or "all"}

\item{ncensus}{The number of the census to be considered}

\item{start}{The first time step to be included}

\item{end}{The last time step to be include}

\item{base}{A character vector with the name of the scenario to be used as 
term of comparison}
}
\value{
A list with four elements:
  \itemize{ 
      \item $extTable.means: A \code{data.frame} (\code{data.table}) with the 
             mean probability of extinction for each year, for each scenario
       \item $extTable.sds: A \code{data.frame} (\code{data.table}) with the 
             standard deviation of the probability of extinction for each Time
             Step, for each scenario 
       \item $cumul.ext.means: A \code{data.frame} (\code{data.table}) with 
             the mean cumulative probability of extinction from the Time Step
             \code{start} to \code{end}, for each scenario
       \item $cumul.ext.sds: A \code{data.frame} (\code{data.table}) with 
             standard deviation of the cumulative probability of extinction 
             from the Time Step \code{start} to \code{end}, for each scenario    
         }
         These results are also saved to disk in two xls files:
   \itemize{ 
      \item $Pext_census:  The mean and standard deviation for each Time
              Step (in two separate tabs)     
       \item $Cumulative_Pext_census:  the mean and standard deviation across 
              the selected time steps (in two separate tabs)
           }
       The census number is appended to the name of the xls file.
}
\description{
Calculates the probability of extinction for each time step and a cumulative 
probability between a specified interval. If the number of the requested 
scenarios is > 1, it is possible to indicate a baseline scenario and \code{Pext}
will also calculate the Strictly Standardised Mean Difference 
  (SSMD, Zhang 2007).
  
By default, \code{Pext} calculates the probability of extinction using the 
column "Population Size", however any column(s) in the census file can be 
passed
with the argument \code{headers}. This allows to set different criteria for extinction 
(e.g. no females are left) or to calculate the probability of extinction for 
several populations at once (if these are captured in different columns in the 
census file, see tutorial for an example). By extension, \code{Pext} can 
therefore be used to calculate the probability of any trait(s) to be zero (or 
non-zero by calculating 1 - Pext).
}
\details{
\code{Pext} takes as input the output of \code{collate.census}, which can be 
passed as R object with \code{data}. If  \code{data=NULL}, then
a fully qualified file name of the .rda file saved by \code{collate.census} 
needs to be passed with the argument \code{rda.in}.
 
When \code{start="min"} (default), it is automatically set to 1. When 
  \code{end="max"} it is set to the maximum Time Step value.

\strong{NOTE} that in this function SSMD is calculated as the difference between 
  the parameter of the base scenario minus the parameter of the alternative
  scenario. This means that a negative SSMD indicates an increase probability
  of extinction of the alternative scenario against the base scenario. See the 
  vignette for further discussion on this point.

It is \strong{also important to note} that if both standard deviations of the 
  probabilities 
  used to calculate the SSMD are zero (this may happen, for example, when the
  populations never go extinct, or always go extinct), it is impossible to 
  calculate the SSMD (the denominator is zero), while if the mean
  probabilities are the same, then SSMD=0 and p-value=0.5.
}
\references{
Zhang, X. D. 2007. A pair of new statistical parameters for quality control
  in RNA interference high-throughput screening assays. Genomics 89:552-561.
}

