% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{prep.data}
\alias{prep.data}
\title{Prepare census data to generate plots}
\usage{
prep.data(i, means, sds, traits, scenarios, rm.T0)
}
\arguments{
\item{i}{The scenario position to extract}

\item{means}{A list of data frames containing the mean values}

\item{sds}{A list of data frames containing the standard deviation values}

\item{traits}{A character vector with the name of the traits to be included}

\item{scenarios}{A character vector with the scenarios' names}

\item{rm.T0}{Whether to remove the first row of data. Relevant when it is 
Time Step '0' (default=TRUE)}
}
\description{
This function is used by other \code{HexSimR}'s functions. It combines mean 
  and standard deviation values in one data.table and add a column with the 
  scenario name. It operates on one scenario at the time.
}
