% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/HexSimR.R
\name{ranges}
\alias{ranges}
\title{Calculates descriptive stats from the HexSim generated report 'ranges'}
\usage{
ranges(rep.ranges = NULL, hx = NULL, events = NULL, start = "min",
  end = "max")
}
\arguments{
\item{rep.ranges}{The fully qualified (i.e. including the path) name of the
report}

\item{hx}{The size of one hexagon in hectares}

\item{events}{A character vector with the name of the events for which the
descriptive statistics needs to be calculated}

\item{start}{The first time step to be included in the summary statistics}

\item{end}{The last time step to be include in the summary statistics}
}
\value{
A data.table with summary statistics, and a .csv with the same data is
  also saved to disk
}
\description{
\code{ranges} also calculates the mean number of groups and the mean size of
  territory in ha and sqkm
}

