% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genepop.div.utilities.R
\name{read.multi.gen}
\alias{read.multi.gen}
\title{Read multiple genepop files}
\usage{
read.multi.gen(path.results = NULL, scenarios = "all", pop.name = NULL)
}
\arguments{
\item{path.results}{The path to the 'Results' folder}

\item{scenarios}{A character vector with the scenarios to be processed or
"all"}

\item{pop.name}{The name of the population}
}
\value{
A list with one element for each Scenario,with withing a list of genind obj
}
\description{
\code{multi.read.gen} is a wrapper for \code{read.one.gen} to read all genepop
files for one or more scenarios
}
\seealso{
gen.dist
}
