% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genepop.div.utilities.R
\name{read.one.gen}
\alias{read.one.gen}
\title{Read one genepop}
\usage{
read.one.gen(fname)
}
\arguments{
\item{fname}{A character vector with the name of the input file including the 
path}
}
\description{
This function takes the output from the function \code{clean.genepop} 
  (or \code{multi.clean.genepop}) as input and uses the R package \code{adegenet}
  to convert the input in a genind object.
}
