% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic.helper.fun.R
\name{read.sds}
\alias{read.sds}
\title{Read standard deviation values from output of collate.census}
\usage{
read.sds(scenario, path.results, ncensus)
}
\arguments{
\item{scenario}{A character vector with the name of the scenario}

\item{path.results}{The path to the 'Results' folder}

\item{ncensus}{The number of the census to be considered}
}
\description{
This function is internally used by other \code{HexSimR} functions
}
