% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rename.replicates.R
\name{rename.replicates}
\alias{rename.replicates}
\title{Rename (renumber) replicate folders within a scenario}
\usage{
rename.replicates(path.results = NULL, scenario, suffix, verbose = FALSE)
}
\arguments{
\item{path.results}{The path to the 'Results' folder}

\item{scenario}{The scenario whose replicates need to be renamed}

\item{suffix}{An integer vector to rename the replicates}

\item{verbose}{Should progress be printed in the console? Default: FALSE}
}
\description{
This function replaces the numbers between square brackets at the end of the
folder names of replicates within one scenario with the numbers provided in
\code{suffix}
}
\details{
\code{rename.replicates} can process one scenario at the time. Processing of
multiple scenarios with the same suffix should be achieved using a loop (e.g.
lapply). \code{suffix} should be of the same length as the number of
replicates.

\bold{Note} that the order in which the original folder names are replaced
may not be progressive. For example, normally the replicate '10' is renamed
after replicate '1' and before '2'.
}
