% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batch.utilities.R
\name{report.batch}
\alias{report.batch}
\title{Write a .xml file to be passed to HexSim to queued 'movements' and/or 'ranges'
 reports}
\usage{
report.batch(path.results = NULL, scenarios = "all", ranges = TRUE,
  move = TRUE)
}
\arguments{
\item{path.results}{The path to the 'Results' folder}

\item{scenarios}{A character vector with the scenarios to be processed or "all"}

\item{ranges}{Whether a batch file to generate 'ranges' reports should be created}

\item{move}{Whether a batch file to generate 'movement' reports should be created}
}
\value{
A .xml file named batchFile_Reports.xml
}
\description{
Generate a batch .xml file in the workspace directory that will instruct 
  OutputTransformer.exe to generate 'movements' and/or 'ranges' 
  report. 
  
It assumes that a combined log file have been generated from all replicates 
  for the given scenarios.
  
If \code{scenarios="all"}, then all the scenarios will be included.
}

