% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HexSimR.R
\name{SSMD.census}
\alias{SSMD.census}
\title{Compare census values against a baseline scenario.}
\usage{
SSMD.census(path.results = NULL, scenarios = "all", base = NULL,
  ncensus = 0)
}
\arguments{
\item{path.results}{The path to the 'Results' folder}

\item{scenarios}{A character vector with the scenarios to be processed or "all"}

\item{base}{A character vector with the name of the scenario to be used as 
term of comparison}

\item{ncensus}{The number of the census to be considered}
}
\value{
A list with SSMD in the first element and p-values in the second. 
  These results are also saved to disk as two tabs in an excel file named 
  "SSMD_census[ncensus].xlsx", where [ncensus] is the number of the census file.
}
\description{
\code{SSMD.census} carries out pairwise comparisons of the census values 
  against a baseline scenario using Strictly Standardised Mean Difference 
  (SSMD, Zhang 2007).  
  
It takes as data input the output from \code{collate.census} (it reads data
  directly from xls files).
}
\references{
Zhang, X. D. 2007. A pair of new statistical parameters for quality control
in RNA interference high-throughput screening assays. Genomics 89:552-561.
}
