% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HexSimR.R
\name{SSMD.move}
\alias{SSMD.move}
\title{Compare mean movement distances against those of a baseline scenario.}
\usage{
SSMD.move(path.results = NULL, scenarios = "all", base = NULL,
  sum.move = "summary_move.csv")
}
\arguments{
\item{path.results}{The path to the 'Results' folder}

\item{scenarios}{A character vector with the scenarios to be processed or "all"}

\item{base}{A character vector with the name of the scenario to be used as 
term of comparison}

\item{sum.move}{The name of the file where the data are (including the extension).}
}
\value{
A list with SSMD in the first element and p-values in the second. 
  These results are also saved to disk as two tabs in an excel file.
}
\description{
\code{SSMD.move} carries out pairwise comparisons of the mean movement distances 
  against those of a baseline scenario using Strictly Standardised Mean Difference 
  (SSMD, Zhang 2007).  
  
It directly reads the .csv files written with \code{move}. It assumes that all 
  files have the same name, which, if different from the default, can be passed
  with the argument\code{sum.move}.
}
\references{
Zhang, X. D. 2007. A pair of new statistical parameters for quality control
in RNA interference high-throughput screening assays. Genomics 89:552-561.
}

