% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HexSimR.R
\name{SSMD.ranges}
\alias{SSMD.ranges}
\title{Compare ranges descriptive statistics against a baseline scenario.}
\usage{
SSMD.ranges(path.results = NULL, scenarios = "all", base = NULL,
  sum.ranges = "summary_ranges.xlsx")
}
\arguments{
\item{path.results}{The path to the 'Results' folder}

\item{scenarios}{A character vector with the scenarios to be processed or "all"}

\item{base}{A character vector with the name of the scenario to be used as 
term of comparison}

\item{sum.ranges}{The name of the file where the data are (including the extension).}
}
\value{
A list with SSMD in the first element and p-values in the second. 
  These results are also saved to disk as two tabs in an excel file.
}
\description{
\code{SSMD.ranges} carries out pairwise comparisons of the descriptive 
statistics calculated froma Ranges report 
  against a baseline scenario using Strictly Standardised Mean Difference 
  (SSMD, Zhang 2007).  
  
It takes as data input the output from \code{ranges} (it reads data
  directly from xls files).
}
\references{
Zhang, X. D. 2007. A pair of new statistical parameters for quality control
in RNA interference high-throughput screening assays. Genomics 89:552-561.
}
