% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/generic.helper.fun.R
\name{ssmd2xlsx}
\alias{ssmd2xlsx}
\title{Save SSMD results to xlsx file}
\usage{
ssmd2xlsx(i, scenarios, ssmds, pvalues, wb)
}
\arguments{
\item{i}{Numeric vector with seq 1 to number of scenarios}

\item{scenarios}{Character vector with names of scenarios}

\item{ssmds}{data.frame with SSMD values}

\item{pvalues}{data.frame with SSMD values}

\item{wb}{Workbook where to save the tabs}
}
\description{
Two tabs are saved for each scenarios: SSMD and p values.
}

