% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/batch.utilities.R
\name{w.combine.log.batch}
\alias{w.combine.log.batch}
\title{Write a .bat file to generate combined log files in batch mode for all
 selected scenarios}
\usage{
w.combine.log.batch(path.results, scenarios = "all", dir.out = NULL)
}
\arguments{
\item{path.results}{The path for the 'Results' folder}

\item{scenarios}{A character vector with scenarios to be processed or "all"}

\item{dir.out}{Path where to save the output. If NULL (default), the working
directory is used}
}
\value{
A .bat file named comb_log_files.bat
}
\description{
Write a .bat file to generate combined log files in batch mode for all
  selected scenarios. When \code{scenarios="all"} all scenarios listed in the
  'Results' folder within the workspace will be included in the batch file.
}
\details{
The file is saved in the path indicated by \code{dir.out}, typically where
  HexSimCommandLine.exe is located. If NULL (default),
  the working directory is used. This may be useful when the batch file is
  being prepared on a different machine. Ultimately, the batch file needs to
  be copied where the HexSimCommandLine.exe is located.

To run the batch file, open the command prompt, navigate where
  HexSimCommandLine.exe is located and type "comb_log_files.bat".
}

