% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batch.utilities.R
\name{w.csvmap.batch}
\alias{w.csvmap.batch}
\title{Write a .bat file to generate hexmap from a list of csv files}
\usage{
w.csvmap.batch(
  file.list,
  header = "true",
  hexID = "true",
  rows,
  cols,
  narrow = "false",
  dir.out = NULL
)
}
\arguments{
\item{file.list}{A character vector with the fully qualified name of the csv 
files}

\item{header, hexID, rows, cols, narrow}{arguments to be passed to 
HexMapConverter.exe, see HexSim guide for details}

\item{dir.out}{Path where to save the output. If NULL (default), the working 
directory is used}
}
\value{
A .bat file named comb_log_files.bat
}
\description{
The .bat file is saved in the path indicated by \code{dir.out}. If NULL 
(default), the working directory is used. Ultimately, the batch file needs to
be copied where the HexMapConverter.exe is located.
}
\details{
The output folder where the hxn are saved will be a subfolder within 
\code{dir.out}. These subfolders are named from the csv file name, without
the extension.

To run the batch file, open the command prompt, navigate where 
HexMapConverter.exe is located and type "csvmap.batch.bat".
}
