% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xml_utilities.R
\name{workspace.path.modifier}
\alias{workspace.path.modifier}
\title{Modify the path of the workspace and grid file in all scenario files}
\usage{
workspace.path.modifier(
  path.scenarios = NULL,
  scenarios = "all",
  new.grid.path
)
}
\arguments{
\item{path.scenarios}{The path to the 'Scenarios' folder}

\item{scenarios}{A character vector with the scenarios to be processed or
"all"}

\item{new.grid.path}{The fully qualified (i.e. including the full path) of 
the grid file, including extension}
}
\description{
When a HexSim workspace is copied to a new location, the path to the root of 
the workspace and the grid file that is stored in the scenario files need to 
be updated. When the simulations are run manually via the GUI, HexSim picks 
this up and asks the user to re-save the scenario. However, when simulations 
are run via command-line, an error occurs unless the path and the grid file 
exist. This function updates all the xml scenario files with the new path
passed with: \code{new.grid.path}.
}
\details{
\bold{WARNING:} Currently, xml scenario files saved by \code{HexSimR} trigger 
a request by HexSim to re-save the files. This is due to very minor 
differences in the file formatting, which do not affect the simulations. 
These files run with no problems when run via the command-line (i.e. 
batchRunner.exe)
}
