% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HexSimR.R
\name{collate.census}
\alias{collate.census}
\title{Combine together HexSim census outputs across iterations.}
\usage{
collate.census(path.results = NULL, scenarios = "all")
}
\arguments{
\item{path.results}{The path to the 'Results' folder}

\item{scenarios}{A character vector with the scenarios to be processed or "all"}
}
\value{
A list with three elements: the combined data, the mean and standard
  deviation. An xls file is also saved with the descriptive statistics and a 
  .rda file is saved with the combined data
}
\description{
Combine together HexSim census outputs across iterations. If the scenarios have 
multiple census events, all census files are processed (separately). Mean values 
are reported in the final file which is located in Results/scenario_name
folder and returned as a list.
}
\details{
If there are several scenarios and \code{scenarios="all"} (default), all
scenarios are processed, otherwise it is possible to select a subset of scenarios
using a character vector, e.g. scenarios=c("scen1", "scen2").

If \code{path.results=NULL} an interactive dialog box is used to select the 
  path where the results are located

Note, when there is a large number of files, this function may be memory hungry
}

