% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{gen.plot}
\alias{gen.plot}
\title{Plot genetic distance with standard deviation bars}
\usage{
gen.plot(path.results = NULL, scenarios = "all", pop.name = NULL,
  time.step = 1, traits, r = 2, c = 1)
}
\arguments{
\item{path.results}{The path to the 'Results' folder}

\item{scenarios}{A character vector with the scenarios to be processed or 
"all"}

\item{pop.name}{The name of the population}

\item{time.step}{A numeric vector of length 1 to indicate the time step to 
be included}

\item{traits}{A character vector with the traits that were included in the 
\code{m.gen.dist} call}

\item{r}{The raw in the genetic distance matrix to use}

\item{c}{The column in the genetic distance matrix to use}
}
\description{
\code{gen.plot} generates a plot of the mean genetic distances with  
  standard deviation bars. It assumes that file names that have been generated
  with \code{m.gen.dist} are as for \code{HexSimR}
  defaults.
}
