% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make.tables.R
\name{make.table}
\alias{make.table}
\title{Make a table from HexSimR results}
\usage{
make.table(path.results = NULL, scenarios = "all", fnames, SSMD = FALSE,
  colh = list(c("GroupSize", "Resources", "nGroups", "ha", "sqkm")),
  vround = 1, sdround = 1, time.steps = NULL,
  table.name = "Tables.xlsx", tab.name = NULL, save2disk = TRUE,
  dir.out = NULL)
}
\arguments{
\item{path.results}{The path to the 'Results' folder}

\item{scenarios}{A character vector with the scenarios to be processed or "all"}

\item{fnames}{The name(s) of the output file(s) to be processed. If more the 
one scenario is passed, the output file names have to be all equal}

\item{SSMD}{Whether the data being processed are SSMD comparisons (default:
FALSE)}

\item{colh}{A list with as many elements as the length of fnames. Each 
element is a character vector with the names of the column headings to be 
considered for the relative fnames (default: ranges' headings)}

\item{vround}{The number of digits the variable (e.g. the mean, the p-values)
should be rounded to}

\item{sdround}{If SSMD=FALSE, the number of digits the standard deviation 
should be rounded to}

\item{time.steps}{A numeric vector with the time steps to include. If NULL
(default), all time steps are included. See details for more information}

\item{table.name}{The name of the file where to save the table}

\item{tab.name}{The name of the tab to be saved in the xls file. By default, 
"Descriptive" or "SSMD", depending on whether \code{SSMD=FALSE} or 
\code{SSMD=TRUE} respectively.}

\item{save2disk}{Whether to save to disk the table}

\item{dir.out}{The directory where to save the table. If NULL (default), the 
file will be written in the 'Results' folder}
}
\value{
A data.frame (data.table) with the compiled table. If save2disk=TRUE 
  an xls file is also saved to disk.
}
\description{
\code{make.table} reads the result files generated by HexSimR and compiles a 
table after having rounded the decimal digits.
}
\details{
\code{make.table} can process either descriptive statistics (generated using 
\code{collate.census}, \code{ranges} or \code{move}) or statistical 
comparisons generated using \code{SSMD.census}, \code{SSMD.ranges} or 
\code{SSMD.move}), but only one type of result in any one time. For each run 
of \code{make.table} a sheet is created as output in a .xls file. If the name
of the xls file is kept equal, the new sheet is added to the existing file.
\strong{ATTENTION:} If the name of the new sheet is the same of an existing
one, the data will be pasted over in the existing sheet, possibly making the
results incomprehensible.

When processing census data, several files can be passed at once by passing a
character vector of length > 1 to \code{fnames}. Note that the extension 
(.csv) of the file is used to identify whether it was requested to process 
descriptive statistics generated by movement reports. If you have changed the
extensions of the result files, this function may have impredictable results.

When more than one scenario is processed, then \code{make.table} assumes 
that, within each scenario, the result file name is the same. The only 
exception is when processing the descriptive statistics of census data. In 
this case, the name of scenario is automatically added to the suffix. In 
these cases, the suffix of the files has to be passed to \code{fnames} (i.e. 
\emph{"[ncensus].all.xlsx"}, where \emph{"ncensus"} is the number of the 
census event, and \emph{".all.xlsx"} is the standard suffix that HexSimR adds
to all files generated with \code{collate.census}.

\code{colh} is used to pass the headings of the columns to be processed. It 
\strong{MUST} be passed as a list, with an element for each file that was 
passed with \code{fnames}.

If \code{SSMD=TRUE} and \code{scenarios="all"}, R may throw an error because 
the base scenario is included in \code{scenarios} (unless it was manually 
removed) but it is not present in the result file. When \code{SSMD=TRUE} a
character vector has to be passed to \code{scenarios}.

When \code{SSMD=TRUE}, currently, only p-values are reported in the table.

Note that when \code{SSMD=TRUE}, and the function is pointed to a SSMD 
comparison of ranges report statistics, the appropriate setting for the time
step is \code{time.steps=NULL} (default) because these are already subsetted
when the SSMD comparison is done, unless all time steps were kept when calling 
\code{ranges} or \code{multi.reports}. Also, unless issue #3 has been resolved,
the argument  \code{time.steps} only works correctly if all time steps were 
logged.

By default, the sheet in the xls file takes name "Descriptive" or "SSMD", 
depending on whether \code{SSMD=FALSE} or \code{SSMD=TRUE} respectively. This
behaviour can be altered by passing a character vector to \code{tab.name}.
}
\seealso{
\code{\link{collate.census}}, \code{\link{ranges}}, 
  \code{\link{move}}, \code{\link{SSMD.census}}, \code{\link{SSMD.ranges}}, 
  \code{\link{SSMD.move}}
}
