% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xml_utilities.R
\name{make.Xpath}
\alias{make.Xpath}
\title{Build a Xpath}
\usage{
make.Xpath(
  node_path,
  identifier,
  attrib,
  param_node = NA,
  param_identifier = FALSE,
  is.LHS = FALSE
)
}
\arguments{
\item{node_path}{The node_path}

\item{identifier}{The identifier of the node}

\item{attrib}{Whether the identifier is an attribute}

\item{param_node}{Whether the parameter is contained in an internal node}

\item{param_identifier}{A character vector if the param has an identifier,
FALSE otherwise}

\item{is.LHS}{Whether the Xpath being built is for a LHS}
}
\value{
A character vector with the Xpath
}
\description{
This function builds a Xpath to search nodes in HexSim xml files. It is used
internally by hexSimR. When used in the LHS.scenarios function, it constructs
the path "/.../name[text()='...']" when param_node is not NA,
param_identifier is FALSE and is.LHS is TRUE. Addtional details on arguments
are in the scenarios.batch.modifier.
}
