% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic.helper.fun.R
\name{pval}
\alias{pval}
\title{Calculate p values}
\usage{
pval(x)
}
\description{
\code{pval} calculates one-tailed p values from a data.frame that contains 
  quantiles and it is used internally from SSMD functions.
  @param x The data.frame of quantiles
}

