% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HexSimR.R
\name{ranges}
\alias{ranges}
\title{Calculates descriptive stats from the HexSim generated report 'ranges'}
\usage{
ranges(rep.ranges = NULL, hx = NULL, events = NULL, start = "min",
  end = "max")
}
\arguments{
\item{rep.ranges}{The fully qualified (i.e. including the path) name of the 
report}

\item{hx}{The size of one hexagon in hectares}

\item{events}{A character vector with the name of the events for which the 
mean and standard deviation across years needs to be calculated. If NULL 
(default), all events are considered}

\item{start}{The first time step to be included}

\item{end}{The last time step to be included}
}
\value{
A list with three elements:
  \itemize{ 
      \item $descriptive: A \code{data.frame} (\code{data.table}) with the 
             summary statistics with mean for each time step, for each event
       \item $means: A \code{data.frame} (\code{data.table}) with the mean 
             across the selected time steps
       \item $sds: A \code{data.frame} (\code{data.table}) with the standard 
             deviation across the selected time steps      
         }
   These results are also saved to disk in two files:
   \itemize{ 
      \item $descriptive_ranges.csv:  The summary statistics for each time 
              step, for each event     
       \item $summary_ranges.xlsx:  the mean and standard deviation across 
              the selected time steps (in two separate tabs)
           }
}
\description{
\code{ranges} calculates descriptive stats, for each year, from the HexSim
generated report 'ranges'. In addition to HexSim items, \code{ranges} also
calculates the mean number of groups and the mean size of territory in ha and
sqkm based on the value passed with \code{hx}. From this table \code{ranges}
then calculates the mean across all years in the window of years 
\code{start} - \code{end}.
}
