% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove.logs.R
\name{remove.logs}
\alias{remove.logs}
\title{Remove log files}
\usage{
remove.logs(
  path.results = NULL,
  scenarios = "all",
  delete.log = TRUE,
  delete.gz = TRUE,
  parallel = FALSE,
  ncores = NULL
)
}
\arguments{
\item{path.results}{The path to the 'Results' folder}

\item{scenarios}{A character vector with the scenarios to be processed or
"all"}

\item{delete.log}{whether the log file should be deleted}

\item{parallel}{whether execute the function in parallel}

\item{ncores}{the number of cores if executed in parallel}
}
\value{
a list where the first element is the list of log files found and the
  second is the gz files created
}
\description{
Remove log files (compressed or as .log) to save space and make
simulation results more portable
}
\details{
This function search for files ending with '.log' or '.gz' in the path provided
and remove them. Care should be taken if there are other files in the directory
that are not the target files.

Note that it makes sense to execute this function in parallel only if a solid
state disk (SSD) is available. If \code{ncores} is left to its default
(\code{NULL}) and \code{parallel=TRUE} all available processors will be used.
}
\examples{
# Create a temp directory
tmp <- tempdir()
dir.create(file.path(tmp, "test"))
# Create a fake log file
cat(file=file.path(tmp, "test", "foo.log"), "Hello world!")
# Compress the log file
remove.logs(path.results=tmp, scenarios="test", delete.log=TRUE)
# clean up
unlink(tmp, recursive=TRUE)
}
