% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/genepop.utilities.R
\name{w.genepop.batch}
\alias{w.genepop.batch}
\title{Write a .xml file to generate genepop input files in batch mode for all
replicates of given scenario(s)}
\usage{
w.genepop.batch(path.results, scenarios = "all", time.steps = 1, pop.name,
  traits)
}
\arguments{
\item{path.results}{The path for the 'Results' folder}

\item{scenarios}{A character vector with scenarios to be processed or "all"}

\item{time.steps}{A numeric vector to indicate the time step to be included}

\item{pop.name}{The name of the population}

\item{traits}{A character vector with the name of the traints to be included}
}
\value{
A .xml file named batchFile_genepop_Reports.xml
}
\description{
Generate a batch .xml file in the workspace directory that will instruct
  OutputTransformer.exe to generate genepop input files for all replicates of
  the scenario(s) passed with \code{scenarios}. If \code{scenarios="all"} is
  used, then all the scenairos will be inlcuded.
}

